Add support for removing the unlisted Flatpak apps (#482)

+++ src/core/flatpaks_installer.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python3
-import os, subprocess, sys, shutil
+import os, subprocess, shutil
 from pathlib import Path
 
 CACHE_FLATPAK = f"{Path.home()}/.var/app/io.github.vikdevelop.SaveDesktop/cache/tmp"
@@ -22,57 +22,54 @@ if os.path.exists(f"{CACHE_FLATPAK}/workspace"):
 else:
     dest_dir = None
    
-# If the destination directory variable is not 'None', continue in installing Flatpak apps
 if dest_dir:
-    # If the destination directory has a directory with installed Flatpak apps user data, install them
+    # Restore Flatpak user data archive if exists
     if os.path.exists(f"{dest_dir}/app"):
         print("copying the Flatpak apps' user data to the ~/.var/app directory")
         os.system(f"cp -au {dest_dir}/app/ ~/.var/")
-    
-    # If the flatpak-apps-data.tgz archive exists, unpack it to the ~/.var/app directory
+
     if os.path.exists(f"{dest_dir}/flatpak-apps-data.tgz"):
         subprocess.run(["tar", "-xzvf", "flatpak-apps-data.tgz", "-C", f"{Path.home()}/.var"])
 
-    # If the Bash scripts for installing Flatpak apps to the system exist, install them
-    if os.path.exists(f"{dest_dir}/installed_flatpaks.sh") or os.path.exists(f"{dest_dir}/installed_user_flatpaks.sh"):
-        print("installing the Flatpak apps on the system")
+    # Load Flatpaks from bash scripts
+    installed_flatpaks_files = [f'{dest_dir}/installed_flatpaks.sh', f'{dest_dir}/installed_user_flatpaks.sh']
+    desired_flatpaks = {}
+    for file in installed_flatpaks_files:
+        if os.path.exists(file):
+            with open(file, 'r') as f:
+                for line in f:
+                    line = line.strip()
+                    if line.startswith('flatpak install'):
+                        parts = line.split()
+                        if len(parts) < 4:
+                            continue
+                        app_id = parts[3]
+                        method = "--user" if "--user" in parts else "--system"
+                        desired_flatpaks[app_id] = method
 
-        # If the Flathub repository in user installation doesn't exist, add it
-        if os.path.getsize(f"{dest_dir}/installed_user_flatpaks.sh") > 5:
-            os.system("flatpak remote-add --user --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo")
-        # Install Flatpak apps from list
-        installed_flatpaks_files = [f'{dest_dir}/installed_flatpaks.sh', f'{dest_dir}/installed_user_flatpaks.sh']
-        system_flatpak_dir = '/var/lib/flatpak/app'
-        user_flatpak_dir = os.path.expanduser('~/.local/share/flatpak/app')
-        # Load Flatpaks from bash scripts
-        installed_flatpaks = {}
-        for file in installed_flatpaks_files:
-            if os.path.exists(file):
-                with open(file, 'r') as f:
-                    for line in f:
-                        if line.startswith('flatpak install'):
-                            parts = line.split()
-                            if '--user' in parts:
-                                app = parts[3]
-                                installed_flatpaks[app] = '--user'
-                            elif '--system' in parts:
-                                app = parts[3]
-                                installed_flatpaks[app] = '--system'
+    # Get currently installed Flatpaks
+    system_flatpak_dir = '/var/lib/flatpak/app'
+    user_flatpak_dir = os.path.expanduser('~/.local/share/flatpak/app')
+    installed_apps = {}
+    for directory, method in [(system_flatpak_dir, '--system'), (user_flatpak_dir, '--user')]:
+        if os.path.exists(directory):
+            for app in os.listdir(directory):
+                if os.path.isdir(os.path.join(directory, app)):
+                    installed_apps[app] = method
 
-        # Get installed Flatpaks in the specified directories
-        installed_apps = set()
-        for directory in [system_flatpak_dir, user_flatpak_dir]:
-            if os.path.exists(directory):
-                installed_apps.update(app for app in os.listdir(directory) if os.path.isdir(os.path.join(directory, app)))
+    # Remove Flatpaks, which are not in the list
+    for app, method in installed_apps.items():
+        if app not in desired_flatpaks:
+            print(f"[REMOVE] {method.title()} Flatpak: {app}")
+            subprocess.run(['flatpak', 'uninstall', method, app, '--delete-data', '-y'])
 
-        # Compare Flatpaks listed in the Bash scripts with the installed ones
-        flatpaks_to_install = {app: method for app, method in installed_flatpaks.items() if app not in installed_apps}
+    for app, method in desired_flatpaks.items():
+        if app not in installed_apps:
+            print(f"[INSTALL] Installing {app} ({method})")
+            if method == '--user':
+                os.system("flatpak remote-add --user --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo")
+            subprocess.run(['flatpak', 'install', method, 'flathub', app, '-y'])
 
-        if flatpaks_to_install:
-            for app, method in flatpaks_to_install.items():
-                subprocess.run(['flatpak', 'install', method, 'flathub', app, '-y'])
-        else:
-            print('All Flatpak apps are installed.')
 else:
     print("Nothing to do.")
 
@@ -82,4 +79,6 @@ if os.path.exists(autostart_file):
     os.remove(autostart_file)
 
 # Remove the cache dir after finishing the operations
-shutil.rmtree(f"{CACHE_FLATPAK}/workspace")
+if os.path.exists(f"{CACHE_FLATPAK}/workspace"):
+    shutil.rmtree(f"{CACHE_FLATPAK}/workspace")
+
